Clean Up Project - Remove Work Items From Project
In this commit, I removed the work items that were encoded in the
project README.md file and added them as issues to the Localizer project
in GitHub.  I also checked the project for TODO items and checked for
comments in the acceptance testing modules since I occasionally track
work items in one of those two ways, too.